pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Ilya <ilya.sereda.2001@gmail.com>"]
readme = "README.md"
license = "MIT"

[tool.black]
line-length = 79
target-version = ["py311"]
include = "\\.pyi?$"
exclude = '''
/(
  \.toml
  |\.git
  |\.ini
  |Dockerfile
)/
'''

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
colorama = "^0.4.6"
telegraph = "^2.2.0"
python-telegram-bot = {extras = ["job-queue", "rate-limiter"], version = "^20.7"}
aiosqlite = "^0.19.0"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pre-commit = "^3.4.0"

[tool.poetry.scripts]
migration = "src.db.migrations.db_init:init_db"
migration-revert = "src.db.migrations.db_delete:delete_db"
seeder = "src.db.seeders.insert_movies:insert"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
